♻️ refactor: CategoryList -> index file naming,  remove console.log
